+++ modules/modules.cpp
@@ -3,6 +3,7 @@
 #include "_template/template.h"
 #include "fish/fish.h"
 #include "flying_toasters/flying_toasters.h"
+#include "mowing_man/mowing_man.h"
 #include "paperboy/paperboy.h"
 #include "starry_night/starry_night.h"
 
@@ -15,12 +16,13 @@ void register_all_modules(Registry& reg) {
   reg.add("flying-toasters", &make_flying_toasters);
   reg.add("starry-night", &make_starry_night);
   reg.add("fish", &make_fish);
+  reg.add("mowing-man", &make_mowing_man);
   // The copy-me template (also a working ambient starfield).
   reg.add("starfield", &make_template_starfield);
 
   // More flagship + classic recreations land here as they're built:
-  //   reg.add("mowing-man", &make_mowing_man);
-  //   reg.add("warp",       &make_warp);
+  //   reg.add("warp",     &make_warp);
+  //   reg.add("spotlight", &make_spotlight);
   //   ...
 }